requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.2.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
cmdstanpy==1.2.5
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.56.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
holidays==0.68
httplib2==0.22.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_resources==6.5.2
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.1.7
nbclient==0.10.2
nbformat==5.10.4
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
prophet==1.1.6
proto-plus==1.25.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
stack-data==0.6.3
stanio==0.5.1
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
